requirements.txt
streamlit
openai
python-dotenv
langchain-core
plotly
pandas